moved mongodb init code; DELETE endpoint
